+++ src/backend/cdb/cdbresynchronizechangetracking.c
@@ -1306,7 +1306,7 @@ ChangeTracking_GetIncrementalChangeList(void)
 	IncrementalChangeList* result = NULL;
 	StringInfoData 	sqlstmt;
 	int 			ret;
-	int 			proc;
+	int 			proc; /* 32 bit, only holds metadata */
 	volatile bool 	connected = false;
 	ResourceOwner 	save = CurrentResourceOwner;
 	MemoryContext 	oldcontext = CurrentMemoryContext;
@@ -1360,7 +1360,7 @@ ChangeTracking_GetIncrementalChangeList(void)
 
 		/* Do the query. */
 		ret = SPI_execute(sqlstmt.data, true, 0);
-		proc = SPI_processed;
+		proc = (int) SPI_processed;
 
 
 		if (ret > 0 && SPI_tuptable != NULL)
@@ -1574,7 +1574,7 @@ ChangeTrackingResult* ChangeTracking_GetChanges(ChangeTrackingRequest *request)
 	ChangeTrackingResult* result = NULL;
 	StringInfoData 	sqlstmt;
 	int 			ret;
-	int 			proc;
+	int 			proc; /* 32 bit, only holds metadata */
 	int				i;
 	volatile bool 	connected = false; /* needs to survive PG_TRY()/CATCH() */
 	ResourceOwner 	save = CurrentResourceOwner;
@@ -1646,7 +1646,7 @@ ChangeTrackingResult* ChangeTracking_GetChanges(ChangeTrackingRequest *request)
 
 		/* Do the query. */
 		ret = SPI_execute(sqlstmt.data, true, 0);
-		proc = SPI_processed;
+		proc = (int) SPI_processed;
 
 
 		if (ret > 0 && SPI_tuptable != NULL)
